+++ async-openai/src/types/admin/users.rs
@@ -17,7 +17,7 @@ pub struct User {
     /// `owner` or `reader`
     pub role: OrganizationRole,
     /// The Unix timestamp (in seconds) of when the users was added.
-    pub added_at: u32,
+    pub added_at: u64,
 }
 
 /// A list of `User` objects.
